Convert Cloudinary signature API from TypeScript to JavaScript
